+++ preprocess/our_data/our_data_creat_csv.py
@@ -16,8 +16,8 @@ def save_our_data_to_csv(image_dir, output_file):
 
 
 def create():
-    output_file_path = os.path.join(os.getcwd(), 'data', 'our_data.csv')
-    image_dir = os.path.join(os.getcwd(), 'data', 'our_data_image from _google')
+    output_file_path = os.path.join(os.getcwd(), '../data', 'our_data.csv')
+    image_dir = os.path.join(os.getcwd(), '../data', 'our_data_image from _google')
     save_our_data_to_csv(image_dir, output_file_path)
