+++ preprocess/view/cifr10_show_image_from_our_csv_path.py
