+++ preprocess/view/show_cifar100_npz.py
